fixes #282 file upload (multiple)
we use phantomjs WebPage#uploadFile to upload multiple files.  we pass
"head" as the selector to WebPage#uploadFile.  this will indirectly
trigger a click event on the head element.

there is no official api so we follow chromedriver and separate files
by the "\n" character.

we return immediatly if one of the files does not exist.